# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_call_service.Call:
    properties:
      callId:
        type: string
      clientId:
        type: string
      createdAt:
        type: string
      direction:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.CallDirection'
      endedAt:
        type: string
      instanceId:
        type: string
      peer:
        type: string
      reason:
        type: string
      status:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.CallStatus'
      updatedAt:
        type: string
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_call_service.RejectCallStruct:
    properties:
      callCreator:
        $ref: '#/definitions/types.JID'
      callId:
        type: string
    required:
    - callCreator
    - callId
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct:
    properties:
      chat:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chat_service.HistorySyncRequestStruct:
    properties:
      count:
        type: integer
      messageInfo:
        $ref: '#/definitions/types.MessageInfo'
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_community_service.CreateCommunityStruct:
    properties:
      communityName:
        type: string
    type: object
...
  pkg_call_handler.startRequest:
    properties:
      number:
        type: string
    required:
    - number
    type: object
  pkg_call_handler.webRTCRequest:
    properties:
      sdpOffer:
        type: string
    required:
    - sdpOffer
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
  /call/{callId}:
    delete:
      parameters:
      - description: Call ID
        in: path
        name: callId
        required: true
        type: string
      - description: Call client identifier
        in: header
        name: X-Call-Client-ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.Call'
      summary: End a call
      tags:
      - Call
  /call/{callId}/accept:
    post:
      parameters:
      - description: Call ID
        in: path
        name: callId
        required: true
        type: string
      - description: Call client identifier
        in: header
        name: X-Call-Client-ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Accept an incoming call
      tags:
      - Call
  /call/{callId}/reject:
    post:
      parameters:
      - description: Call ID
        in: path
        name: callId
        required: true
        type: string
      - description: Call client identifier
        in: header
        name: X-Call-Client-ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.Call'
      summary: Reject an incoming call
      tags:
      - Call
  /call/{callId}/webrtc:
    post:
      consumes:
      - application/json
      parameters:
      - description: Call ID
        in: path
        name: callId
        required: true
        type: string
      - description: Call client identifier
        in: header
        name: X-Call-Client-ID
        required: true
        type: string
      - description: SDP offer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_call_handler.webRTCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Negotiate call WebRTC
      tags:
      - Call
  /call/active:
    get:
      parameters:
      - description: Optional call client filter
        in: header
        name: X-Call-Client-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List active calls
      tags:
      - Call
  /call/events:
    get:
      parameters:
      - description: Optional call client filter
        in: header
        name: X-Call-Client-ID
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Stream call events
      tags:
      - Call
  /call/reject:
    post:
      consumes:
      - application/json
      description: Reject a WhatsApp call using callCreator and callId.
      parameters:
      - description: Call data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.RejectCallStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reject call (legacy)
      tags:
      - Call
  /call/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: Call client identifier
        in: header
        name: X-Call-Client-ID
        required: true
        type: string
      - description: Destination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_call_handler.startRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/gin.H'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/gin.H'
      summary: Start a WhatsApp voice call
      tags:
      - Call
  /chat/archive:
    post:
      consumes:
      - application/json
      description: Archive a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Archive a chat
      tags:
      - Chat
  /chat/history-sync:
    post:
      consumes:
      - application/json
      description: HistorySyncRequest a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.HistorySyncRequestStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: HistorySyncRequest a chat
      tags:
      - Chat
  /chat/mute:
    post:
      consumes:
      - application/json
      description: Mute a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Mute a chat
      tags:
      - Chat
  /chat/pin:
    post:
      consumes:
      - application/json
      description: Pin a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Pin a chat
      tags:
      - Chat
  /chat/unarchive:
    post:
      consumes:
      - application/json
      description: Unarchive a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unarchive a chat
      tags:
      - Chat
  /chat/unmute:
    post:
      consumes:
      - application/json
      description: Unmute a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unmute a chat
      tags:
      - Chat
  /chat/unpin:
    post:
      consumes:
      - application/json
      description: Unpin a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unpin a chat
      tags:
      - Chat
  /community/add:
    post:
      consumes:
      - application/json
      description: Add participant to community
      parameters:
      - description: Participant data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_community_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Add participant to community
      tags:
      - Community
  /community/create:
    post:
      consumes:
      - application/json
      description: Create community
      parameters:
      - description: Community data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_community_service.CreateCommunityStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create community
      tags:
      - Community
  /community/remove:
    post:
      consumes:
      - application/json
      description: Remove participant from community
      parameters:
      - description: Participant data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_community_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Remove participant from community
      tags:
      - Community
  /group/create:
    post:
      consumes:
      - application/json
      description: Create group
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.CreateGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create group
      tags:
      - Group
  /group/description:
    post:
      consumes:
      - application/json
      description: Set group description
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.SetGroupDescriptionStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set group description
      tags:
      - Group
  /group/info:
    post:
      consumes:
      - application/json
      description: Get group info
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.GetGroupInfoStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get group info
      tags:
      - Group
  /group/invitelink:
    post:
      consumes:
      - application/json
      description: Get group invite link
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.GetGroupInviteLinkStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get group invite link
      tags:
      - Group
  /group/join:
    post:
      consumes:
      - application/json
      description: Join group link
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.JoinGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Join group link
      tags:
      - Group
  /group/leave:
    post:
      consumes:
      - application/json
      description: Leave group
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.LeaveGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Leave group
      tags:
      - Group
  /group/list:
    get:
      consumes:
      - application/json
      description: List groups
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List groups
      tags:
      - Group
  /group/myall:
    get:
      consumes:
      - application/json
      description: Get my groups
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get my groups
      tags:
      - Group
  /group/name:
    post:
      consumes:
      - application/json
      description: Set group name
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.SetGroupNameStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set group name
      tags:
      - Group
  /group/participant:
    post:
      consumes:
      - application/json
      description: Update participant
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update participant
      tags:
      - Group
...
swagger: "2.0"